resolved crash bug

+++ app/src/main/java/com/digitaldukaan/models/response/LeadsResponse.kt
@@ -8,7 +8,7 @@ data class LeadsResponse(
     @SerializedName("customer_name")            var customerName: String = "",
     @SerializedName("phone_number")             var phoneNumber: String = "",
     @SerializedName("last_updated_on")          var lastUpdateOn: String = "",
-    @SerializedName("order_value")              var orderValue: Int = 0,
+    @SerializedName("order_value")              var orderValue: Double = 0.0,
     @SerializedName("notification_sent_on")     var notificationSentOn: String = "",
     @SerializedName("cart_type")                var cartType: Int = 0,
     var updatedDate: Date? = Date()
